add searchbar component
