+++ src/components/Home/Searchbar.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Searchbar = ({ onChange }) => {
+  return (
+    <input
+      type="text"
+      name="searchbar"
+      id="sbAlphaCode"
+      placeholder="Top level domain"
+      maxLength="2"
+      onChange={onChange}
+    />
+  );
+};
+
+export default Searchbar;
